Rename file with json schemas
